fix optimizer tasks constructions (#4261)
Conflicts:
	ydb/core/tx/columnshard/engines/storage/optimizer/lbuckets/optimizer.h
	ydb/core/tx/columnshard/hooks/testing/controller.h
	ydb/core/tx/tiering/ut/ut_tiers.cpp
